# catotel-api/openapi/catotel-api.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/users/register": {
      "post": {
        "operationId": "UserController_register",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterUserDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserProfileDto"
                }
              }
            }
          }
        },
        "summary": "Register a new Catotel account",
        "tags": [
          "Users"
        ]
      }
    },
    "/users/me": {
      "get": {
        "operationId": "UserController_getMe",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserProfileDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Return the authenticated user profile",
        "tags": [
          "Users"
        ]
      }
    },
    "/customers/me": {
      "get": {
        "operationId": "CustomersController_getMe",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "tags": [
          "Customers"
        ]
      },
      "patch": {
        "operationId": "CustomersController_updateProfile",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateCustomerDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "tags": [
          "Customers"
        ]
      }
    },
    "/customers/cats": {
      "get": {
        "operationId": "CustomersController_listCats",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "tags": [
          "Customers"
        ]
      },
      "post": {
        "operationId": "CustomersController_createCat",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCatDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "tags": [
          "Customers"
        ]
      }
    },
    "/customers/cats/{id}": {
      "patch": {
        "operationId": "CustomersController_updateCat",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateCatDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "tags": [
          "Customers"
        ]
      }
    },
    "/rooms": {
      "get": {
        "operationId": "RoomsController_list",
        "parameters": [
          {
            "name": "includeInactive",
            "required": false,
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Rooms"
        ]
      },
      "post": {
        "operationId": "RoomsController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateRoomDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "tags": [
          "Rooms"
        ]
      }
    },
    "/rooms/{id}": {
      "patch": {
        "operationId": "RoomsController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateRoomDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "tags": [
          "Rooms"
        ]
      }
    },
    "/reservations": {
      "get": {
        "operationId": "ReservationsController_list",
        "parameters": [
          {
            "name": "status",
            "required": false,
            "in": "query",
            "schema": {
              "enum": [
                "PENDING",
                "CONFIRMED",
                "CHECKED_IN",
                "CHECKED_OUT",
                "CANCELLED"
              ],
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "tags": [
          "Reservations"
        ]
      },
      "post": {
        "operationId": "ReservationsController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateReservationDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "tags": [
          "Reservations"
        ]
      }
    },
    "/reservations/{id}": {
      "get": {
        "operationId": "ReservationsController_getById",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "tags": [
          "Reservations"
        ]
      }
    },
...
  },
  "info": {
    "title": "Catotel API",
    "description": "Auth-first API for the Catotel platform. Includes device sessions, token rotation, and health checks.",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "access-token": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "description": "Provide the short-lived access token returned by /auth/login"
      }
    },
    "schemas": {
...
      "RegisterUserDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "example": "newcat@catotel.com"
          },
          "password": {
            "type": "string",
            "minLength": 8,
            "example": "Purrfect123"
          },
          "name": {
            "type": "string",
            "example": "Luna",
            "nullable": true,
            "description": "Optional display name to show in dashboards"
          }
        },
        "required": [
          "email",
          "password"
        ]
      },
      "CustomerSummaryDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "phone": {
            "type": "object"
          },
          "address": {
            "type": "object"
          }
        },
        "required": [
          "id"
        ]
      },
      "StaffSummaryDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "position": {
            "type": "object"
          },
          "phone": {
            "type": "object"
          }
        },
        "required": [
          "id"
        ]
      },
      "UserProfileDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "email": {
            "type": "string",
            "example": "newcat@catotel.com"
          },
          "name": {
            "type": "object",
            "example": "Luna",
            "nullable": true
          },
          "role": {
            "type": "string",
            "enum": [
              "ADMIN",
              "MANAGER",
              "STAFF",
              "CUSTOMER"
            ],
            "example": "CUSTOMER"
          },
          "customer": {
            "$ref": "#/components/schemas/CustomerSummaryDto"
          },
          "staff": {
            "$ref": "#/components/schemas/StaffSummaryDto"
          }
        },
        "required": [
          "id",
          "email",
          "role"
        ]
      },
      "UpdateCustomerDto": {
        "type": "object",
        "properties": {
          "phone": {
            "type": "string"
          },
          "preferredVet": {
            "type": "string"
          },
          "address": {
            "type": "string"
          },
          "emergencyContactName": {
            "type": "string"
          },
          "emergencyContactPhone": {
            "type": "string"
          },
          "notes": {
            "type": "string"
          }
        }
      },
      "CreateCatDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Mila"
          },
          "breed": {
            "type": "string"
          },
          "gender": {
            "type": "string",
            "enum": [
              "FEMALE",
              "MALE",
              "UNKNOWN"
            ]
          },
          "birthDate": {
            "type": "string"
          },
          "weightKg": {
            "type": "number",
            "description": "Weight in kg"
          },
          "dietaryNotes": {
            "type": "string"
          },
          "medicalNotes": {
            "type": "string"
          },
          "photoUrl": {
            "type": "string"
          }
        },
        "required": [
          "name"
        ]
      },
      "UpdateCatDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Mila"
          },
          "breed": {
            "type": "string"
          },
          "gender": {
            "type": "string",
            "enum": [
              "FEMALE",
              "MALE",
              "UNKNOWN"
            ]
          },
          "birthDate": {
            "type": "string"
          },
          "weightKg": {
            "type": "number",
            "description": "Weight in kg"
          },
          "dietaryNotes": {
            "type": "string"
          },
          "medicalNotes": {
            "type": "string"
          },
          "photoUrl": {
            "type": "string"
          }
        }
      },
      "CreateRoomDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "capacity": {
            "type": "number",
            "default": 1
          },
          "nightlyRate": {
            "type": "number",
            "description": "Nightly rate in USD"
          },
          "amenities": {
            "type": "object",
            "description": "JSON string of amenities"
          },
          "isActive": {
            "type": "boolean"
          }
        },
        "required": [
          "name",
          "capacity",
          "nightlyRate"
        ]
      },
      "UpdateRoomDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "capacity": {
            "type": "number",
            "default": 1
          },
          "nightlyRate": {
            "type": "number",
            "description": "Nightly rate in USD"
          },
          "amenities": {
            "type": "object",
            "description": "JSON string of amenities"
          },
          "isActive": {
            "type": "boolean"
          }
        }
      },
      "CreateReservationDto": {
        "type": "object",
        "properties": {}
      },
...
    }
  }
}